+++ src/components/antivirus/scan/scan-process.tsx
@@ -1,72 +1,73 @@
 import Popup from "@/components/popup";
 import { Button } from "@/components/ui/button";
-import { ButtonGroup } from "@/components/ui/button-group";
 import { Progress } from "@/components/ui/progress";
 import { Spinner } from "@/components/ui/spinner";
 import { SCAN_TYPES } from "@/lib/constants";
 import { ScanType } from "@/lib/types";
-import { cn } from "@/lib/utils";
-import { Bug, Clock, Folder, Pause, Search, Square, Timer } from "lucide-react";
-import { useEffect, useMemo, useState } from "react";
-import { invoke } from "@tauri-apps/api/core";
+import { Bug, Clock, Dot, Folder, SearchCheck, ShieldAlert, ShieldCheck, Square } from "lucide-react";
+import { useMemo, useRef, useState } from "react";
 
 interface Props{
      scanType: ScanType,
-     onStop: () => void
+     onStop: () => void,
+     threatsCount: number,
+     currLocation: string,
+     filesCount: number,
+     totalFiles: number
 }
-export default function ScanProcess({scanType, onStop}: Props){
+
+export default function ScanProcess({scanType, onStop, threatsCount, currLocation, filesCount, totalFiles}: Props){
      const scanTypeName = useMemo(()=>SCAN_TYPES.find(val=>val.type===scanType)?.name,[scanType])
-     if(scanType==="") return null;
      const [isOpen, setIsOpen] = useState(false);
-     const [isPaused, setIsPaused] = useState(false);
+     const dateRef = useRef<Date>(new Date(Date.now()))
      const handleStopScan = () => {
           onStop();
           setIsOpen(false);
      }
-     useEffect(()=>{
-          (async()=>{
-               try{
-                    await invoke(`start_${scanType}_scan`)
-               } catch {
-                    console.error("Command not found")
-               }
-          })()
-     },[])
+     const percentage = useMemo(()=>totalFiles>0 ? Math.min(100,Math.floor((filesCount/totalFiles)*100)) : 0,[filesCount,totalFiles])
      return (
           <>
                <p className="text-muted-foreground font-medium">Scan Type: {scanTypeName}</p>
-               <Progress value={50}/>
-               <p className="text-2xl font-semibold text-center flex justify-center items-center gap-3"><Spinner className={cn("size-9 text-muted-foreground",isPaused && "paused")}/> 50% - 12345 files scanned</p>
-               <ButtonGroup className="w-full">
-                    <Button className="flex-1" onClick={()=>setIsPaused(!isPaused)}>
-                         {isPaused ? <Search/> : <Pause/>}
-                         {isPaused ? "Resume" : "Pause"} the Scan
-                    </Button>
-                    <Button variant="secondary" className="flex-1" onClick={()=>setIsOpen(true)}><Square/> Stop the Scan</Button>
-               </ButtonGroup>
+               {scanType!=="full" ? (
+                    <>
+                    <Progress value={percentage}/>
+                    <p className="text-2xl font-semibold text-center flex justify-center items-center gap-3"><Spinner className="size-9 text-muted-foreground"/> {filesCount<=0 ? "Preparing Scan..." : `${percentage}% - ${filesCount}/${totalFiles} files scanned`}</p>
+                    </>
+               ) : (
+                    <>
+                    <p className="text-2xl font-semibold text-center flex justify-center items-center gap-1.5"><Spinner className="size-9 text-muted-foreground"/> {filesCount<=0 ? "Preparing scan…" : (
+                         <>
+                              Scanning... <Dot/> {filesCount} files scanned
+                         </>
+                    )}</p>
+                    <p className="text-muted-foreground">Full scans may take a while depending on disk size.</p>
+                    </>
+               )}
                <div className="flex justify-center items-center gap-2 flex-col">
-                    <div className="p-4 border bg-card text-card-foreground shadow-sm rounded-md w-full">
-                         <h2 className="text-lg sm:text-xl font-semibold flex items-center gap-2.5 border-b pb-0.5 mb-2 w-fit"><Folder className="text-primary"/> Current Location</h2>
-                         <code>C:\Users\User\...</code>
-                    </div>
+                    {(scanType==="custom" || scanType==="file") && (
+                         <div className="p-4 border bg-card text-card-foreground shadow-sm rounded-md w-full">
+                              <h2 className="text-lg sm:text-xl font-semibold flex items-center gap-2.5 border-b pb-0.5 mb-2 w-fit"><Folder className="text-primary"/> Scan Location</h2>
+                              <code className="break-all">C:\Users\User\...</code>
+                         </div>
+                    )}
+                    {currLocation.trim() !== "" && (
+                         <div className="p-4 border bg-card text-card-foreground shadow-sm rounded-md w-full">
+                              <h2 className="text-lg sm:text-xl font-semibold flex items-center gap-2.5 border-b pb-0.5 mb-2 w-fit"><SearchCheck className="text-primary"/> Currently scanned</h2>
+                              <code className="min-h-6 break-all">{currLocation}</code>
+                         </div>
+                    )}
                     <div className="p-4 border bg-card text-card-foreground shadow-sm rounded-md w-full">
                          <h2 className="text-lg sm:text-xl font-semibold flex items-center gap-2.5 border-b pb-0.5 mb-2 w-fit"><Bug className="text-primary"/> Findings</h2>
-                         <p>2 Threats</p>
+                         <p className="flex items-center gap-2 min-h-6">{threatsCount<=0 ? <ShieldCheck className="text-emerald-700"/> : <ShieldAlert className="text-destructive"/>} {threatsCount} Threats</p>
                     </div>
                     <div className="p-4 border bg-card text-card-foreground shadow-sm rounded-md w-full">
-                         <h2 className="text-lg sm:text-xl font-semibold flex items-center gap-2.5 border-b pb-0.5 mb-2 w-fit"><Folder className="text-primary"/> Scan Location</h2>
-                         <code>C:\Users\User\...</code>
+                         <h2 className="text-lg sm:text-xl font-semibold flex items-center gap-2.5 border-b pb-0.5 mb-2 w-fit"><Clock className="text-primary"/> Start Time</h2>
+                         <code className="min-h-6">{dateRef.current.toLocaleString()}</code>
                     </div>
                     <div className="grid grid-cols-1 lg:grid-cols-2 gap-2 w-full">
-                         <div className="p-4 border bg-card text-card-foreground shadow-sm rounded-md w-full">
-                              <h2 className="text-lg sm:text-xl font-semibold flex items-center gap-2.5 border-b pb-0.5 mb-2 w-fit"><Clock className="text-primary"/> Start Time</h2>
-                              <code>05-05-2020 05:05PM</code>
-                         </div>
-                         <div className="p-4 border bg-card text-card-foreground shadow-sm rounded-md w-full">
-                              <h2 className="text-lg sm:text-xl font-semibold flex items-center gap-2.5 border-b pb-0.5 mb-2 w-fit"><Timer className="text-primary"/> Duration</h2>
-                              <code>00:00:30</code>
-                         </div>
+                         
                     </div>
+                    <Button className="flex-1" onClick={()=>setIsOpen(true)}><Square/> Stop the Scan</Button>
                </div>
                <Popup
                     open={isOpen}